move optimize to a subpackage
